Refactor : 테스트 코드 정리
