+++ backend/src/domain/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module, forwardRef } from "@nestjs/common";
 import { UserModule } from "src/domain/user/user.module";
-import { AuthController } from "./controller/auth.controller";
-import { AuthService } from "./service/auth.service";
+import { AuthController } from "./auth.controller";
+import { AuthService } from "./auth.service";
 import { NaverStrategy } from "./strategy/naver.strategy";
 import { JwtModule } from "@nestjs/jwt";
 import { JwtStrategy } from "./strategy/jwt.strategy";
