+++ backend/src/domain/auth/guard/jwt-auth-guard.ts
@@ -1,6 +1,6 @@
 import { ExecutionContext, Injectable } from "@nestjs/common";
 import { AuthGuard } from "@nestjs/passport";
-import { AuthService } from "../service/auth.service";
+import { AuthService } from "../auth.service";
 
 @Injectable()
 export class JwtAuthGuard extends AuthGuard("jwt") {
