--- backend/src/domain/group/controller/group.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { GroupController } from "./group.controller";
-
-describe("GroupController", () => {
-  let controller: GroupController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [GroupController],
-    }).compile();
-
-    controller = module.get<GroupController>(GroupController);
-  });
-
-  it("should be defined", () => {
-    expect(controller).toBeDefined();
-  });
-});
