--- backend/src/domain/post/controller/post.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { PostController } from "./post.controller";
-
-describe("PostController", () => {
-  let controller: PostController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [PostController],
-    }).compile();
-
-    controller = module.get<PostController>(PostController);
-  });
-
-  it("should be defined", () => {
-    expect(controller).toBeDefined();
-  });
-});
