+++ backend/src/domain/post/post.service.ts
@@ -1,8 +1,8 @@
 import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from "@nestjs/common";
-import { ImageService } from "src/domain/image/service/image.service";
+import { ImageService } from "src/domain/image/image.service";
 import { getImagesUrl } from "src/common/imageUrl";
 import { InjectRepository } from "@nestjs/typeorm";
-import { PostRepository } from "../post.repository";
+import { PostRepository } from "./post.repository";
 import { UserRepository } from "src/domain/user/user.repository";
 import { AlbumRepository } from "src/domain/album/album.repository";
 import { CreatePostRequestDto } from "src/dto/post/createPostRequest.dto";
@@ -10,8 +10,8 @@ import { GetPostInfoResponseDto } from "src/dto/post/getPostInfoResponse.dto";
 import { UpdatePostInfoRequestDto } from "src/dto/post/updatePostInfoRequest.dto";
 import { ShiftPostRequestDto } from "src/dto/post/shiftPostRequest.dto";
 import { GetSearchPostResponse } from "src/dto/post/getSearchPostResponse.dto";
-import { AlbumService } from "src/domain/album/service/album.service";
-import { HashTagService } from "src/domain/hashtag/service/hashtag.service";
+import { AlbumService } from "src/domain/album/album.service";
+import { HashTagService } from "src/domain/hashtag/hashtag.service";
 import { HashTag } from "src/domain/hashtag/hashtag.entity";
 import { Post } from "src/domain/post/post.entity";
 import { HashTagRepository } from "src/domain/hashtag/hashtag.repository";
