+++ backend/src/domain/post/spec/post.service.spec.ts
@@ -5,14 +5,14 @@ import { HashTagRepository } from "src/domain/hashtag/hashtag.repository";
 import { UserRepository } from "src/domain/user/user.repository";
 import { Connection } from "typeorm";
 import { PostRepository } from "../post.repository";
-import { PostService } from "./post.service";
+import { PostService } from "../post.service";
 import { Post } from "../post.entity";
 import { User } from "src/domain/user/user.entity";
 import { Image } from "src/domain/image/image.entity";
 import { GetPostInfoResponseDto } from "src/dto/post/getPostInfoResponse.dto";
-import { AlbumService } from "src/domain/album/service/album.service";
-import { ImageService } from "src/domain/image/service/image.service";
-import { HashTagService } from "src/domain/hashtag/service/hashtag.service";
+import { AlbumService } from "src/domain/album/album.service";
+import { ImageService } from "src/domain/image/image.service";
+import { HashTagService } from "src/domain/hashtag/hashtag.service";
 import { GroupRepository } from "src/domain/group/group.repository";
 import { ImageRepository } from "src/domain/image/image.repository";
 import { Album } from "src/domain/album/album.entity";
