--- backend/src/domain/user/controller/user.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { UserController } from "./user.controller";
-
-describe("UserController", () => {
-  let controller: UserController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [UserController],
-    }).compile();
-
-    controller = module.get<UserController>(UserController);
-  });
-
-  it("should be defined", () => {
-    expect(controller).toBeDefined();
-  });
-});
