+++ backend/src/domain/user/user.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { UserController } from "./controller/user.controller";
-import { UserService } from "./service/user.service";
+import { UserController } from "./user.controller";
+import { UserService } from "./user.service";
 import { UserRepository } from "./user.repository";
 import { JwtModule } from "@nestjs/jwt";
 import { AuthModule } from "src/domain/auth/auth.module";
